Started moving fetching to react query
